+++ 0x15-file_io/main.h
@@ -13,7 +13,7 @@
  * Functions Prototype
  **********************/
 ssize_t read_textfile(const char *filename, size_t letters);
-size_t _strlen(char *s);
+int _strlen(char *s);
 int create_file(const char *filename, char *text_content);
 
 #endif /* MAIN_H */
